Rate limiting is now done by limiting number of gooutines
First, a channel is created called sem (as it will act as a semaphore)
with the level of concurrency desired. As we loop over the urls, we
attempt to put a bool onto the channel. If it isn't full, we fire off
the goroutine on the URL, which defers a read from the semaphore which
frees its slot.