Move to glide for vendoring
This way we move to go "standard" vendoring practices
including using the vendor/ dir in a way compatible
with the go toolchain.

As this is a final binary, we use the --strip-vcs and --v option
in glide up to flatten the dependency tree and vendor the source as is.

This allows this repository to be self sufficient which is necessary atm for
our jenkins based C.I.

The result is therefore comparable to what we had with gb.