Fix: Closing the Body of the response after err checking
We need to close the Body of the response to ensure that the
ressources are freed and that we can re-use the TCP connection
for subsequent HTTP requests.
These ressources are held by the HTTPClient in the gopĥercloud.Provider
so they can be shared by any http request we make.
If we don't close the body of the http.Response then we will create a
new TCP connection for each http request which is very wasteful.

NOTA BENE: we need to close AFTER checking the request didn't err !
gopherCloud will close the Body itself in case of an error.

TODO:  We might want to have a copy of gopherCloud.Provider per worker
I'm still not very clear on the performance implication of sharing this ressource.
But that is for another day... Premature optimisation is the root of all Evil !